pyproject.toml
[tool.poetry]
name = "task-python"
version = "0.1.0"
description = ""
authors = ["Anton Shefer <rezajkee@mail.ru>"]
readme = "README.md"
packages = [{include = "task_python"}]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.89.1"
uvicorn = {extras = ["standart"], version = "^0.20.0"}
SQLAlchemy = "1.4.46"
alembic = "^1.9.2"
databases = "^0.7.0"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.5"
python-decouple = "^3.7"
asyncclick = "^8.1.3.4"
PyJWT = "^2.6.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.11.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
extend-exclude = '''
/(
  | migrations
)/
'''
